requirements.txt
gradio
torch
torchvision
opencv-python
Pillow
numpy
